+++ lib/ask/web_fetch.rb
@@ -1,4 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'web_fetch/version'
+require_relative 'web_fetch/backend'
+require_relative 'web_fetch/backends/local'
+require_relative 'web_fetch/backends/jina'
 require_relative 'web_fetch/tool'